+++ packages/i18nify-js/src/index.ts
@@ -3,3 +3,4 @@ export * from './modules/currency';
 export * from './modules/phoneNumber';
 export * from './modules/dateTime';
 export * from './modules/geo';
+export * from './modules/types';